requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# AI/LLM
anthropic==0.34.0
sentence-transformers==2.3.1
chromadb==0.4.22

# PDF Processing
pypdf==4.2.0
Pillow==10.3.0

# LangChain - minimal for text splitting only
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langgraph==0.1.0